Update images.js
Hi, there is a bug in Openstack where you cannot set the metadata of os_type as it is overwritten to None when first creating the image. This causes a problem with the timezone not being set correctly on windows machines. So I wrote a function, updateImageMeta, that will set the value of os_type to windows after calling createImage function and wanted to share it so maybe it will be in future version of pkgcloud and help other people who might need this functionality as well.